Function dipstick::stats_all

pub fn stats_all(
    kind: InputKind, 
    name: MetricName, 
    score: ScoreType
) -> Option<(InputKind, MetricName, MetricValue)>

A predefined export strategy reporting all aggregated stats for all metric types. Resulting stats are named by appending a short suffix to each metric's name.