Struct dipstick::NameParts

pub struct NameParts { /* fields omitted */ }

A double-ended vec of strings constituting a metric name or a future part thereof.

Implementations

impl NameParts

pub fn is_within(&self, other: &NameParts) -> bool

Returns true if this instance is equal to or a subset (more specific) of the target instance. e.g. a.b.c is within a.b e.g. a.d.c is not within a.b

pub fn make_name<S: Into<String>>(&self, leaf: S) -> MetricName

Make a name in this namespace

pub fn short(&self) -> MetricName

Extract a copy of the last name part Panics if empty