Trait dipstick::Cancel

pub trait Cancel {
    fn cancel(&self);

    fn into_guard(self) -> CancelGuard<Self>
    where
        Self: Sized,
    { ... }
}

A deferred, repeatable, background action that can be cancelled.

Required methods

fn cancel(&self)

Cancel the action.

Loading content...
Provided methods

fn into_guard(self) -> CancelGuard<Self> where
Self: Sized,

Create a guard that cancels when it is dropped.

Loading content...